Fixed security bug related to PID's and psutil
A session without "psutil" would only store the PID but not the
executable name which would result in a new session with "psutil"
available to ignore the name-checking, causing a security problem if the
PID would be occupied by another process. The user could potentially
kill a unrelated process resulting in unwanted side-effects.